+++ solvers/schrodinger_solver.py
@@ -30,7 +30,7 @@ def initial_condition(x: torch.Tensor) -> torch.Tensor:
     """Exact IC h(x, 0) = 2*sech(x) (real part; imaginary part 0). Also the
     whole-domain target of the PirateNets physics-informed output init
     (u(x,t) ≈ u0(x) for all t). Returns (N, 2) = [Re, Im]."""
-    real = (2.0 / torch.cosh(x[:, :1])).float()
+    real = 2.0 / torch.cosh(x[:, :1])
     return torch.cat([real, torch.zeros_like(real)], dim=1)
 
 
@@ -239,7 +239,9 @@ def generate_dataset(
     N = n_residual + n_ic + n_bc
     x = torch.zeros(N, spatial_dim, device=device)
     t = torch.zeros(N, 1, device=device)
-    h_gt = torch.zeros(N, 2, device=device, dtype=torch.float32)
+    # x/t/h_gt follow torch's default dtype, set from config 'precision'
+    # before generation — GT targets keep float64 accuracy in double runs.
+    h_gt = torch.zeros(N, 2, device=device)
 
     x_min, x_max = spatial_domain[0]
     t_min, t_max = temporal_domain
@@ -250,16 +252,16 @@ def generate_dataset(
     print(f"  Sampling {n_residual} residual points from grid...")
     i_t = np.random.choice(nt, size=n_residual, replace=True)
     i_x = np.random.choice(nx, size=n_residual, replace=True)
-    x[idx:idx + n_residual, 0] = torch.from_numpy(x_grid[i_x].astype(np.float32)).to(device)
-    t[idx:idx + n_residual, 0] = torch.from_numpy(t_grid[i_t].astype(np.float32)).to(device)
-    h_gt[idx:idx + n_residual, 0] = torch.from_numpy(h_solution[i_t, i_x].real.astype(np.float32)).to(device)
-    h_gt[idx:idx + n_residual, 1] = torch.from_numpy(h_solution[i_t, i_x].imag.astype(np.float32)).to(device)
+    x[idx:idx + n_residual, 0] = torch.from_numpy(x_grid[i_x]).to(device)
+    t[idx:idx + n_residual, 0] = torch.from_numpy(t_grid[i_t]).to(device)
+    h_gt[idx:idx + n_residual, 0] = torch.from_numpy(h_solution[i_t, i_x].real.astype(np.float64)).to(device)
+    h_gt[idx:idx + n_residual, 1] = torch.from_numpy(h_solution[i_t, i_x].imag.astype(np.float64)).to(device)
     idx += n_residual
 
     # IC: sample random x from grid, t=t_min
     print(f"  Sampling {n_ic} initial condition points from grid...")
     i_x_ic = np.random.choice(nx, size=n_ic, replace=True)
-    x[idx:idx + n_ic, 0] = torch.from_numpy(x_grid[i_x_ic].astype(np.float32)).to(device)
+    x[idx:idx + n_ic, 0] = torch.from_numpy(x_grid[i_x_ic]).to(device)
     t[idx:idx + n_ic, 0] = t_min
     idx += n_ic
 
@@ -270,17 +272,17 @@ def generate_dataset(
     i_t_bc = np.random.choice(nt, size=max(n_bc_left, n_bc_right), replace=True)
 
     x[idx:idx + n_bc_left, 0] = x_min
-    t[idx:idx + n_bc_left, 0] = torch.from_numpy(t_grid[i_t_bc[:n_bc_left]].astype(np.float32)).to(device)
+    t[idx:idx + n_bc_left, 0] = torch.from_numpy(t_grid[i_t_bc[:n_bc_left]]).to(device)
     # BC h_gt from solver: h_solution at (i_t, x=x_min) which is index 0
-    h_gt[idx:idx + n_bc_left, 0] = torch.from_numpy(h_solution[i_t_bc[:n_bc_left], 0].real.astype(np.float32)).to(device)
-    h_gt[idx:idx + n_bc_left, 1] = torch.from_numpy(h_solution[i_t_bc[:n_bc_left], 0].imag.astype(np.float32)).to(device)
+    h_gt[idx:idx + n_bc_left, 0] = torch.from_numpy(h_solution[i_t_bc[:n_bc_left], 0].real.astype(np.float64)).to(device)
+    h_gt[idx:idx + n_bc_left, 1] = torch.from_numpy(h_solution[i_t_bc[:n_bc_left], 0].imag.astype(np.float64)).to(device)
     idx += n_bc_left
 
     x[idx:idx + n_bc_right, 0] = x_max
-    t[idx:idx + n_bc_right, 0] = torch.from_numpy(t_grid[i_t_bc[:n_bc_right]].astype(np.float32)).to(device)
+    t[idx:idx + n_bc_right, 0] = torch.from_numpy(t_grid[i_t_bc[:n_bc_right]]).to(device)
     # For periodic grid x_max is not in the grid; use first index (periodicity)
-    h_gt[idx:idx + n_bc_right, 0] = torch.from_numpy(h_solution[i_t_bc[:n_bc_right], 0].real.astype(np.float32)).to(device)
-    h_gt[idx:idx + n_bc_right, 1] = torch.from_numpy(h_solution[i_t_bc[:n_bc_right], 0].imag.astype(np.float32)).to(device)
+    h_gt[idx:idx + n_bc_right, 0] = torch.from_numpy(h_solution[i_t_bc[:n_bc_right], 0].real.astype(np.float64)).to(device)
+    h_gt[idx:idx + n_bc_right, 1] = torch.from_numpy(h_solution[i_t_bc[:n_bc_right], 0].imag.astype(np.float64)).to(device)
 
     mask_residual = torch.zeros(N, dtype=torch.bool, device=device)
     mask_residual[:n_residual] = True
@@ -290,7 +292,7 @@ def generate_dataset(
     mask_bc[n_residual + n_ic:] = True
 
     # Overwrite IC with exact analytical values
-    h_gt[mask_ic, 0] = (2.0 / torch.cosh(x[mask_ic, 0])).float()
+    h_gt[mask_ic, 0] = 2.0 / torch.cosh(x[mask_ic, 0])
     h_gt[mask_ic, 1] = 0.0
 
     print("  Dataset generated successfully")
@@ -327,8 +329,8 @@ def evaluate_on_grid(x_grid: torch.Tensor, config: Dict) -> torch.Tensor:
     
     # Extract complex values and split into real/imag
     h_complex = h_solution[i_t, i_x]
-    h_gt = np.zeros((len(x_query), 2), dtype=np.float32)
-    h_gt[:, 0] = h_complex.real.astype(np.float32)
-    h_gt[:, 1] = h_complex.imag.astype(np.float32)
+    h_gt = np.zeros((len(x_query), 2), dtype=np.float64)
+    h_gt[:, 0] = h_complex.real.astype(np.float64)
+    h_gt[:, 1] = h_complex.imag.astype(np.float64)
     
     return torch.from_numpy(h_gt)
